+++ fairscape_models/activity.py
@@ -1,10 +1,10 @@
-from pydantic import BaseModel, Field, ConfigDict, model_validator
-from typing import Optional, List, Union
+from pydantic import Field, ConfigDict, field_validator
+from typing import Optional, List, Union, Any
 
-from fairscape_models.fairscape_base import IdentifierValue
+from fairscape_models.fairscape_base import IdentifierValue, Identifier, extractGUID
 from fairscape_models._version import __version__
 
-class Activity(BaseModel):
+class Activity(Identifier):
     """Base class for Activity types (Computation, Annotation, Experiment)"""
     guid: str = Field(alias="@id")
     name: str
@@ -20,3 +20,9 @@ class Activity(BaseModel):
     fairscapeVersion: str = __version__
 
     model_config = ConfigDict(extra="allow", populate_by_name=True)
+
+#   TODO extract guids from all subfields
+#    @field_validator('generated', mode='before')
+#    @classmethod
+#    def extract_guid_generated(cls, value: Any) -> Any:
+#        pass